refactor(product-detail): move into pages folder
