+++ src/app/pages/product/product-routing.module.ts
@@ -2,9 +2,11 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { ProductListComponent } from './product-list/product-list.component';
+import { ProductDetailComponent } from './product-detail/product-detail.component';
 
 const routes: Routes = [
   { path: '', pathMatch: 'full', component: ProductListComponent },
+  { path: ':id', component: ProductDetailComponent },
 ];
 
 @NgModule({
