--- src/app/product-detail/product-detail.component.scss
@@ -1,38 +1,0 @@
-@import "../color.scss";
-.container {
-    font-family: Microsoft JhengHei;
-    border:2px solid map-get($colors, main);
-    border-radius: 20px;
-    margin: 5% 0;
-    padding: 1rem;
-}
-
-.productDetail{
-    margin-bottom: 30px;
-    .productDetail-img{
-        text-align: center;
-        padding: 1rem;
-    }
-}
-.borrow{
-    text-align: right;
-    .borrow-btn{
-        background:map-get($colors, iris-blue);
-        color:map-get($colors, font-white);
-    }
-}
-
-
-.image{
-    width: 100%;
-    // height: 200px;
-    border-radius: 23px;
-    margin: 0 auto;
-}
-
-
-.expansion-panel{
-    border-radius: 0;
-    width: 90%;
-    margin: 20px auto;
-}
